Move dingo.gw.pipe to dingo.pipe.
