+++ dingo/pipe/ACKNOWLEDGMENT.md
