+++ dingo/pipe/data_generation.py
@@ -3,12 +3,12 @@ import sys
 
 from bilby_pipe.input import Input
 from bilby_pipe.main import parse_args
-from bilby_pipe.utils import log_version_information, logger, convert_string_to_dict
+from bilby_pipe.utils import logger, convert_string_to_dict
 from bilby_pipe.data_generation import DataGenerationInput as BilbyDataGenerationInput
 
 from dingo.gw.data.event_dataset import EventDataset
 from dingo.gw.domains import FrequencyDomain
-from dingo.gw.pipe.parser import create_parser
+from dingo.pipe.parser import create_parser
 
 
 class DataGenerationInput(BilbyDataGenerationInput):
