+++ dingo/pipe/default_settings.py
