+++ dingo/pipe/dingo_result.py
