+++ dingo/pipe/utils.py
