+++ pyproject.toml
@@ -91,11 +91,11 @@ dingo_generate_dataset_dag = "dingo.gw.dataset.generate_dataset_dag:main"
 dingo_ls = "dingo.gw.ls_cli:ls"
 dingo_merge_datasets = "dingo.gw.dataset.utils:merge_datasets_cli"
 dingo_merge_asd_datasets = "dingo.gw.noise.utils:merge_datasets_cli"
-dingo_pipe = "dingo.gw.pipe.main:main"
-dingo_pipe_importance_sampling = "dingo.gw.pipe.importance_sampling:main"
-dingo_pipe_generation = "dingo.gw.pipe.data_generation:main"
-dingo_pipe_sampling = "dingo.gw.pipe.sampling:main"
-dingo_pipe_plot = "dingo.gw.pipe.plot:main"
+dingo_pipe = "dingo.pipe.main:main"
+dingo_pipe_importance_sampling = "dingo.pipe.importance_sampling:main"
+dingo_pipe_generation = "dingo.pipe.data_generation:main"
+dingo_pipe_sampling = "dingo.pipe.sampling:main"
+dingo_pipe_plot = "dingo.pipe.plot:main"
 dingo_train = "dingo.gw.training:train_local"
 dingo_train_condor = "dingo.gw.training.train_pipeline_condor:train_condor"
 dingo_result = "dingo.gw.pipe.dingo_result:main"
